Fix APSTUD-4707 Write out which file failed to index   - right-size collections when we have a good ballpark number of possible size   - For some build participants initialize problems to Collections.emptyList and then assign generated list if we get results (for parse errors, use CollectionsUtil.map to convert the IParseErrors to IProblems)   - Add filename to error logging

+++ plugins/com.aptana.buildpath.core/src/com/aptana/core/internal/build/IndexBuildParticipant.java
@@ -84,6 +84,13 @@ public class IndexBuildParticipant extends RequiredBuildParticipant
 				}
 			}
 
+			// Check for cancellation
+			if (sub.isCanceled())
+			{
+				return;
+			}
+
+			// wipe the index for the file first
 			deleteFile(context, sub.newChild(10));
 
 			List<IFileStoreIndexingParticipant> indexers = getIndexParticipants(context);
